Restructured STDIN handling.
The STDIN handling state is not private and defined in the new include
file REND_STDIN.INS.PAS.  This include file is only referenced by the base
REND_STDIN module, and the system-dependent REND_STDIN_SYS module.

A REND_STDIN_SYS module was added to the WIN32 implementation, but it only
contains dummy routines for now.  STDIN events are still not possible.