Fixed how STDIN-reading thread is stopped.
The correct way to stop the thread is to call CancelIoEx.  However, this
routine is not available in WinXP and earlier.  To allow for the same EXE
to run from WinXP on, CancelIoEx is not called.  All possible events are
signalled, and a reasonable time is given for the thread to exit and
signal that it is done.  After this time (200 ms in this version), the
routine to stop STDIN reading returns anyway.  This is OK for programs
that enable STDIN reading once, then generally exit RENDlib right before
exiting themselves.  It can cause memory leaks and other problems if a
program switches STDIN reading on/off several times, and expects this to
happen quickly.